# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Workshop OS API",
        "description": "API para gerenciamento de ordens de serviço de oficina mecânica. Perfis: receptionist, mechanic, client.",
        "contact": {
            "email": "dev@workshop.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/v1/parts": {
            "get": {
                "tags": [
                    "Parts"
                ],
                "summary": "Lista peças",
                "operationId": "f19a0688d27324fb398104f0fc461d7b",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "low_stock",
                        "in": "query",
                        "description": "Filtra estoque baixo (<=5)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de peças"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Parts"
                ],
                "summary": "Cadastra peça",
                "operationId": "d48eb3679fae100136251b4d0aaa3f79",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Filtro de óleo"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "part_number": {
                                        "type": "string",
                                        "example": "FO-001"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 35
                                    },
                                    "stock_quantity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Peça criada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/parts/{id}": {
            "get": {
                "tags": [
                    "Parts"
                ],
                "summary": "Exibe peça",
                "operationId": "1f29d8130c1a8781cd037611a161aaa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da peça"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Parts"
                ],
                "summary": "Atualiza peça",
                "operationId": "fa044eea40faba7f8c051a5893489c3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "stock_quantity": {
                                        "type": "integer"
                                    },
                                    "active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Peça atualizada"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Parts"
                ],
                "summary": "Remove peça",
                "operationId": "b1435c6188201380f7e4ab3aa0e8f5bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Peça removida"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Informe o token Sanctum: Bearer {token}"
            }
        }
    },
...
}